Organize celery config and tasks.
