+++ manifest_base.yml
@@ -15,11 +15,11 @@ applications:
   instances: 1
   memory: 256M
   no-route: true
-  command: eval $(python webservices/setenv.py) && celery beat --app cron
+  command: celery beat --app webservices.tasks
   buildpack: https://github.com/cloudfoundry/python-buildpack#v1.5.1
 - name: celery-worker
   instances: 1
   memory: 256M
   no-route: true
-  command: eval $(python webservices/setenv.py) && celery worker --app cron
+  command: celery worker --app webservices.tasks
   buildpack: https://github.com/cloudfoundry/python-buildpack#v1.5.1
