+++ webservices/tasks/refresh.py
@@ -1,0 +1,25 @@
+import io
+import logging
+
+import manage
+from webservices import mail
+from webservices.tasks import app
+
+logger = logging.getLogger(__name__)
+
+@app.task
+def refresh():
+    """Update incremental aggregates and materialized views, then email logs
+    to the development team.
+    """
+    buffer = io.StringIO()
+    with mail.CaptureLogs(manage.logger, buffer):
+        try:
+            manage.update_aggregates()
+            manage.refresh_materialized()
+        except Exception as error:
+            manage.logger.exception(error)
+    try:
+        mail.send_mail(buffer)
+    except Exception as error:
+        logger.exception(error)
